if a file is already started or crawled ignore it

+++ dmoz/crawl-file.sh
@@ -2,5 +2,10 @@
 export SITES=$(pwd)/$1
 OUT=${SITES}.gz
 ERR=$(pwd)/$1.err
-scrapy crawl dmoz-csv --loglevel ERROR --output-format=jsonlines --output - 2>${ERR} | pv | gzip > ${OUT}
-gsutil cp ${OUT} gs://websites-classify/raw/
+if [ -f ${OUT} ]; then
+    echo "file ${OUT} already done"
+    exit 0
+else
+    scrapy crawl dmoz-csv --loglevel ERROR --output-format=jsonlines --output - 2>${ERR} | pv | gzip > ${OUT}
+    gsutil cp ${OUT} gs://websites-classify/raw/
+fi
